@@ -1,9 +1,7 @@
-// @ts-check
+import type { NextConfig } from "next";
 
-/**
- * @type {import('next').NextConfig}
- **/
-const nextConfig = {
+const nextConfig: NextConfig = {
+  /* config options here */
   compiler: {
     // Remove `console.*` output except `console.error`
     removeConsole: {
@@ -14,4 +12,4 @@ const nextConfig = {
   },
 };
 
-module.exports = nextConfig;
+export default nextConfig;
